requirements.txt
openai
pandas
tiktoken
psycopg2
pgvector
python-dotenv
pypdf
langchain
supabase